Return RDF::Literals instead of their objects
This is an option for dealing with languages and typed literals.
Literals are returned as an RDF::Literal instead of an (e.g.) String,
Date, or other datatype. This means the type and language of the literal
are accessible after property values have been returned. The underlying
Ruby object can be accessed with #object.

Note that this is a work in progress and fails a number of tests. Some
failures would be API changes for RDF usersin the event this code makes
it into master.